fix: gate HA recovery arm and retain storage on notify drain timeout
Require AdmissionReady with publish-then-recheck before arming recovery
human-attention delivery so BeginShutdown cannot miss a nil cancel and
leave a live notify after stop has drained. Treat stopHumanAttention
WaitGroup timeout as shutdownDrainErr so Stop retains SQLite under a
still-live osascript. Split oversized gateway and recovery contract
tests into focused files under the 300-line subsystem threshold.

Generated-By: looper 0.11.4 (runner=fixer, agent=grok-build)
