--- internal/runtime/human_attention_recovery_contract_test.go
@@ -1,484 +1,0 @@
-package runtime
-
-import (
-	"context"
-	"os"
-	"path/filepath"
-	"strings"
-	"sync"
-	"testing"
-	"time"
-
-	"github.com/nexu-io/looper/internal/config"
-	"github.com/nexu-io/looper/internal/eventlog"
-	networkclient "github.com/nexu-io/looper/internal/network/client"
-	"github.com/nexu-io/looper/internal/storage"
-)
-
-// Cross-component: after durable parks exist without a live finalize callback
-// (crash-before-notify window), recovery rescan emits once; re-rescan dedupes.
-func TestHumanAttentionContract_RecoveryRescanAwaitingAndManual(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-	root := t.TempDir()
-	capturePath := filepath.Join(root, "osascript.log")
-	scriptPath := filepath.Join(root, "osascript")
-	writeHumanAttentionOsascript(t, scriptPath, capturePath, false)
-
-	coordinator := openMigratedCoordinator(t, filepath.Join(root, "recovery-rescan.sqlite"), filepath.Join(root, "backups"))
-	repos := storage.NewRepositories(coordinator.DB())
-	now := time.Date(2026, time.July, 29, 15, 0, 0, 0, time.UTC)
-	nowISO := eventlog.FormatJavaScriptISOString(now)
-
-	projectID := "project_recovery_rescan"
-	if err := repos.Projects.Upsert(ctx, storage.ProjectRecord{
-		ID: projectID, Name: "Recovery Rescan", RepoPath: filepath.Join(root, "repo"),
-		CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Projects.Upsert() error = %v", err)
-	}
-
-	// Durable awaiting_human park with no prior notification (missed finalize callback).
-	awaitLoopID := "loop_await_missed_notify"
-	awaitTarget := projectID
-	if err := repos.Loops.Upsert(ctx, storage.LoopRecord{
-		ID: awaitLoopID, Seq: 701, ProjectID: projectID, Type: "worker",
-		TargetType: "project", TargetID: &awaitTarget, Repo: stringPtr("acme/looper"),
-		Status: "awaiting_human", CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Loops.Upsert(awaiting_human) error = %v", err)
-	}
-	if err := repos.Runs.Upsert(ctx, storage.RunRecord{
-		ID: "run_await_missed", LoopID: awaitLoopID, Status: "running",
-		StartedAt: nowISO, CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Runs.Upsert(await) error = %v", err)
-	}
-
-	// Hard manual_intervention park (latest queue row).
-	manualLoopID := "loop_manual_missed_notify"
-	manualTarget := "pr:acme/looper:99"
-	prNumber := int64(99)
-	if err := repos.Loops.Upsert(ctx, storage.LoopRecord{
-		ID: manualLoopID, Seq: 702, ProjectID: projectID, Type: "fixer",
-		TargetType: "pull_request", TargetID: &manualTarget, Repo: stringPtr("acme/looper"),
-		PRNumber: &prNumber, Status: "paused", CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Loops.Upsert(manual) error = %v", err)
-	}
-	checkpoint := `{"resumePolicy":"manual_intervention"}`
-	if err := repos.Runs.Upsert(ctx, storage.RunRecord{
-		ID: "run_manual_missed", LoopID: manualLoopID, Status: "failed",
-		CheckpointJSON: &checkpoint, Summary: stringPtr("orphan quarantine"),
-		StartedAt: nowISO, EndedAt: &nowISO, CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Runs.Upsert(manual) error = %v", err)
-	}
-	manualKind := "manual_intervention"
-	finished := nowISO
-	if err := repos.Queue.Upsert(ctx, storage.QueueItemRecord{
-		ID: "queue_manual_missed", ProjectID: &projectID, LoopID: &manualLoopID,
-		Type: "fixer", TargetType: "pull_request", TargetID: manualTarget,
-		Repo: stringPtr("acme/looper"), PRNumber: &prNumber,
-		DedupeKey: "fixer:manual_missed", Priority: storage.QueuePriorityFixer,
-		Status: "manual_intervention", AvailableAt: nowISO, Attempts: 1, MaxAttempts: 3,
-		LastError: stringPtr("orphan quarantine"), LastErrorKind: &manualKind,
-		FinishedAt: &finished, CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Queue.Upsert(manual) error = %v", err)
-	}
-
-	// Ordinary exhausted retry under manual_intervention status must stay silent.
-	ordinaryLoopID := "loop_ordinary_missed"
-	ordinaryTarget := projectID
-	if err := repos.Loops.Upsert(ctx, storage.LoopRecord{
-		ID: ordinaryLoopID, Seq: 703, ProjectID: projectID, Type: "worker",
-		TargetType: "project", TargetID: &ordinaryTarget, Status: "paused",
-		CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Loops.Upsert(ordinary) error = %v", err)
-	}
-	ordinaryKind := "retryable_transient"
-	if err := repos.Queue.Upsert(ctx, storage.QueueItemRecord{
-		ID: "queue_ordinary_missed", ProjectID: &projectID, LoopID: &ordinaryLoopID,
-		Type: "worker", TargetType: "project", TargetID: ordinaryTarget,
-		DedupeKey: "worker:ordinary_missed", Priority: storage.QueuePriorityWorker,
-		Status: "manual_intervention", AvailableAt: nowISO, Attempts: 3, MaxAttempts: 3,
-		LastError: stringPtr("network blip"), LastErrorKind: &ordinaryKind,
-		FinishedAt: &finished, CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Queue.Upsert(ordinary) error = %v", err)
-	}
-
-	cfg, err := config.DefaultConfig(root)
-	if err != nil {
-		t.Fatalf("DefaultConfig() error = %v", err)
-	}
-	cfg.Notifications.InApp = true
-	cfg.Notifications.Osascript = config.OsascriptNotificationConfig{
-		Enabled:               true,
-		SoundForLevels:        []config.NotificationSoundLevel{config.NotificationSoundLevelActionRequired},
-		ThrottleWindowSeconds: 60,
-	}
-	osascriptPath := scriptPath
-	cfg.Tools.OsascriptPath = &osascriptPath
-	cfg.Daemon.LogDir = filepath.Join(root, "logs")
-	cfg.Server.AuthMode = config.AuthModeNone
-
-	rt := New(Options{Config: cfg, Logger: &testLogger{}, Now: func() time.Time { return now }})
-
-	ids := collectHumanAttentionLoopIDs(ctx, repos)
-	if len(ids) < 2 {
-		t.Fatalf("collectHumanAttentionLoopIDs() = %v, want at least awaiting + manual", ids)
-	}
-
-	// Synchronous rescan path (same body as the async recovery schedule).
-	rt.notifyDurableHumanAttentionParksBestEffort(ctx, repos)
-	assertHumanAttentionInAppCount(t, repos, awaitLoopID, 1)
-	assertHumanAttentionInAppCount(t, repos, manualLoopID, 1)
-	assertHumanAttentionInAppCount(t, repos, ordinaryLoopID, 0)
-
-	// Permanent entry dedupe: second rescan does not resend.
-	rt.notifyDurableHumanAttentionParksBestEffort(ctx, repos)
-	assertHumanAttentionInAppCount(t, repos, awaitLoopID, 1)
-	assertHumanAttentionInAppCount(t, repos, manualLoopID, 1)
-}
-
-// Recovery human-attention rescan must not launch until startup is committed.
-// If a later CompleteStartup step fails (e.g. network manager), Start closes
-// SQLite — a still-running rescan would query/persist through a closed DB.
-func TestHumanAttentionContract_RecoveryNotifyOnlyAfterStartupCommitted(t *testing.T) {
-	t.Parallel()
-
-	workingDir := t.TempDir()
-	cfg, err := config.DefaultConfig(workingDir)
-	if err != nil {
-		t.Fatalf("DefaultConfig() error = %v", err)
-	}
-	cfg.Storage.DBPath = filepath.Join(workingDir, "runtime.sqlite")
-	backupDir := filepath.Join(workingDir, "backups")
-	cfg.Storage.BackupDir = &backupDir
-	cfg.Daemon.LogDir = filepath.Join(workingDir, "logs")
-	cfg.Notifications.InApp = true
-	cfg.Notifications.Osascript.Enabled = false
-	// Coding agent configured so network manager is exercised after recovery.
-	vendor := config.AgentVendorOpenCode
-	cfg.Agent.Vendor = &vendor
-
-	coordinator, err := storage.OpenSQLiteCoordinator(context.Background(), cfg.Storage.DBPath, storage.SQLiteCoordinatorOptions{
-		BackupDir:  backupDir,
-		Migrations: storage.EmbeddedMigrations,
-	})
-	if err != nil {
-		t.Fatalf("OpenSQLiteCoordinator() error = %v", err)
-	}
-	defer func() { _ = coordinator.Close() }()
-	if _, err := coordinator.MigrationRunner().RunPending(context.Background()); err != nil {
-		t.Fatalf("MigrationRunner().RunPending() error = %v", err)
-	}
-	repositories := storage.NewRepositories(coordinator.DB())
-	startedAt := time.Date(2026, time.July, 29, 17, 0, 0, 0, time.UTC)
-	// Malformed network state makes networkManager.Start fail after recovery.
-	statePath := filepath.Join(workingDir, ".looper", "network.json")
-	if err := os.MkdirAll(filepath.Dir(statePath), 0o755); err != nil {
-		t.Fatalf("MkdirAll() error = %v", err)
-	}
-	if err := os.WriteFile(statePath, []byte("{"), 0o644); err != nil {
-		t.Fatalf("WriteFile() error = %v", err)
-	}
-
-	// Seed a durable park so a wrongly-early schedule would have work to do.
-	nowISO := eventlog.FormatJavaScriptISOString(startedAt)
-	projectID := "project_boot_order"
-	if err := repositories.Projects.Upsert(context.Background(), storage.ProjectRecord{
-		ID: projectID, Name: "Boot Order", RepoPath: filepath.Join(workingDir, "repo"),
-		CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Projects.Upsert() error = %v", err)
-	}
-	loopID := "loop_boot_order_await"
-	target := projectID
-	if err := repositories.Loops.Upsert(context.Background(), storage.LoopRecord{
-		ID: loopID, Seq: 801, ProjectID: projectID, Type: "worker",
-		TargetType: "project", TargetID: &target, Status: "awaiting_human",
-		CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Loops.Upsert() error = %v", err)
-	}
-
-	rt := &Runtime{
-		config:           cfg,
-		logger:           &testLogger{},
-		now:              func() time.Time { return startedAt },
-		shutdownTimeout:  2 * time.Second,
-		startedAt:        &startedAt,
-		services:         Services{Coordinator: coordinator, Repositories: repositories},
-		networkManager:   networkclient.NewManager(statePath, cfg, repositories, nil),
-		admission:        NewAdmission(),
-		startupReadyOnce: sync.Once{},
-	}
-
-	err = rt.CompleteStartup(context.Background())
-	if err == nil {
-		t.Fatal("CompleteStartup() error = nil, want network manager start failure")
-	}
-	if !strings.Contains(err.Error(), "decode network state") {
-		t.Fatalf("CompleteStartup() error = %q, want malformed network state error", err)
-	}
-
-	rt.mu.RLock()
-	cancel := rt.humanAttentionNotifyCancel
-	done := rt.humanAttentionNotifyDone
-	rt.mu.RUnlock()
-	if cancel != nil || done != nil {
-		t.Fatalf("human-attention recovery notify scheduled on failed startup (cancel=%v done=%v); must launch only after MarkReady", cancel != nil, done != nil)
-	}
-	// No notification rows: rescan must not have run against still-open SQLite
-	// before the failure returned (Start would then close the coordinator).
-	assertHumanAttentionInAppCount(t, repositories, loopID, 0)
-}
-
-// Post-claim human-attention delivery must not use the detached WithoutCancel
-// dispatch context: BeginShutdown cancels a runtime-owned parent, and Stop
-// drains so a 35s osascript cannot outlive the shutdown timeout and persist
-// through a closed coordinator.
-func TestHumanAttentionContract_PostClaimNotifyCanceledOnShutdown(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-	root := t.TempDir()
-	startedPath := filepath.Join(root, "osascript.started")
-	scriptPath := filepath.Join(root, "osascript")
-	// Block until SIGTERM/SIGINT (shell.Run kill on ctx cancel). Without cancel
-	// this would sleep ~30s like a real dialog and outlive shutdownTimeout.
-	script := "#!/bin/sh\n: > \"" + startedPath + "\"\ntrap 'exit 0' TERM INT\nwhile true; do sleep 0.05; done\n"
-	if err := os.WriteFile(scriptPath, []byte(script), 0o755); err != nil {
-		t.Fatalf("WriteFile(osascript) error = %v", err)
-	}
-
-	coordinator := openMigratedCoordinator(t, filepath.Join(root, "post-claim-cancel.sqlite"), filepath.Join(root, "backups"))
-	repos := storage.NewRepositories(coordinator.DB())
-	now := time.Date(2026, time.July, 29, 18, 0, 0, 0, time.UTC)
-	nowISO := eventlog.FormatJavaScriptISOString(now)
-	projectID := "project_post_claim_cancel"
-	if err := repos.Projects.Upsert(ctx, storage.ProjectRecord{
-		ID: projectID, Name: "Post Claim Cancel", RepoPath: filepath.Join(root, "repo"),
-		CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Projects.Upsert() error = %v", err)
-	}
-	loopID := "loop_post_claim_cancel"
-	target := projectID
-	if err := repos.Loops.Upsert(ctx, storage.LoopRecord{
-		ID: loopID, Seq: 901, ProjectID: projectID, Type: "worker",
-		TargetType: "project", TargetID: &target, Status: "awaiting_human",
-		CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Loops.Upsert() error = %v", err)
-	}
-
-	cfg, err := config.DefaultConfig(root)
-	if err != nil {
-		t.Fatalf("DefaultConfig() error = %v", err)
-	}
-	cfg.Notifications.InApp = true
-	cfg.Notifications.Osascript.Enabled = true
-	cfg.Notifications.Osascript.ThrottleWindowSeconds = 60
-	osascriptPath := scriptPath
-	cfg.Tools.OsascriptPath = &osascriptPath
-	cfg.Daemon.LogDir = filepath.Join(root, "logs")
-
-	admission := NewAdmission()
-	if err := admission.MarkReady("test post-claim cancel"); err != nil {
-		t.Fatalf("MarkReady() error = %v", err)
-	}
-	rt := &Runtime{
-		config:           cfg,
-		logger:           &testLogger{},
-		now:              func() time.Time { return now },
-		shutdownTimeout:  1500 * time.Millisecond,
-		services:         Services{Coordinator: coordinator, Repositories: repos},
-		admission:        admission,
-		activeExecutions: NewActiveExecutionRegistry(),
-		shutdownCh:       make(chan struct{}),
-	}
-
-	// Simulate post-finalize callback under detached dispatch ctx (WithoutCancel).
-	// Runtime must ignore that and schedule under its cancelable parent.
-	dispatchCtx := context.WithoutCancel(context.Background())
-	rt.notifyHumanAttentionPostClaim(dispatchCtx, loopID)
-
-	// Wait until the fake osascript has started so we know cancel has work to do.
-	deadline := time.Now().Add(3 * time.Second)
-	for {
-		if _, err := os.Stat(startedPath); err == nil {
-			break
-		}
-		if time.Now().After(deadline) {
-			t.Fatal("osascript did not start; post-claim notify may not have launched")
-		}
-		time.Sleep(20 * time.Millisecond)
-	}
-
-	// BeginShutdown cancels HA delivery; stop drains the WaitGroup.
-	start := time.Now()
-	rt.BeginShutdown("test cancel post-claim human-attention")
-	rt.stopHumanAttentionRecoveryNotify()
-	elapsed := time.Since(start)
-	if elapsed > 3*time.Second {
-		t.Fatalf("stopHumanAttentionRecoveryNotify() elapsed = %v, want cancel-bounded (not full osascript dialog)", elapsed)
-	}
-
-	// After cancel+drain, SQLite close must not race a still-running persist.
-	if err := coordinator.Close(); err != nil {
-		t.Fatalf("coordinator.Close() after drained HA notify error = %v", err)
-	}
-}
-
-// Post-recovery rescan is cancel/done-tracked and joined on Stop so SQLite
-// close cannot race the background query (CI TempDir state/ cleanup failures).
-func TestHumanAttentionContract_RecoveryNotifyJoinedBeforeStop(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-	root := t.TempDir()
-	dbPath := filepath.Join(root, "state", "join-stop.sqlite")
-	if err := os.MkdirAll(filepath.Dir(dbPath), 0o755); err != nil {
-		t.Fatalf("MkdirAll(state) error = %v", err)
-	}
-	cfg, err := config.DefaultConfig(root)
-	if err != nil {
-		t.Fatalf("DefaultConfig() error = %v", err)
-	}
-	backupDir := filepath.Join(root, "backups")
-	cfg.Storage.DBPath = dbPath
-	cfg.Storage.BackupDir = &backupDir
-	cfg.Daemon.LogDir = filepath.Join(root, "logs")
-	cfg.Daemon.WorkingDirectory = root
-	cfg.Notifications.InApp = true
-	cfg.Notifications.Osascript.Enabled = false
-	// No coding agent → scheduler may wait, but recovery rescan still schedules.
-	now := time.Date(2026, time.July, 29, 16, 0, 0, 0, time.UTC)
-	rt := New(Options{
-		Config: cfg,
-		Logger: &testLogger{},
-		Now:    func() time.Time { return now },
-		RunSchedulerTick: func(context.Context, Services) error {
-			return nil
-		},
-		ShutdownTimeout: 2 * time.Second,
-	})
-	if err := rt.Start(ctx); err != nil {
-		t.Fatalf("Start() error = %v", err)
-	}
-	if err := rt.CompleteStartup(ctx); err != nil {
-		t.Fatalf("CompleteStartup() error = %v", err)
-	}
-	// Wait path used by API fixtures after CompleteStartup.
-	if err := rt.WaitForHumanAttentionRecoveryNotify(ctx); err != nil {
-		t.Fatalf("WaitForHumanAttentionRecoveryNotify() error = %v", err)
-	}
-	// Stop must join (or cancel+join) before closing SQLite; no hang / panic.
-	rt.Stop("test join human-attention recovery notify")
-	// Second wait is a no-op once the done channel was cleared.
-	if err := rt.WaitForHumanAttentionRecoveryNotify(ctx); err != nil {
-		t.Fatalf("WaitForHumanAttentionRecoveryNotify() after Stop = %v", err)
-	}
-}
-
-// Recovery quarantine must not deliver human-attention notifications on the
-// critical path (interactive dialogs can take tens of seconds per loop).
-func TestHumanAttentionContract_QuarantineDoesNotNotifySynchronously(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-	root := t.TempDir()
-	// Slow osascript would hang startup if quarantine still notified inline.
-	slowScript := filepath.Join(root, "osascript")
-	if err := os.WriteFile(slowScript, []byte("#!/bin/sh\nsleep 30\n"), 0o755); err != nil {
-		t.Fatalf("WriteFile(slow osascript) error = %v", err)
-	}
-
-	coordinator := openMigratedCoordinator(t, filepath.Join(root, "quarantine-sync.sqlite"), filepath.Join(root, "backups"))
-	repos := storage.NewRepositories(coordinator.DB())
-	now := time.Date(2026, time.July, 29, 15, 30, 0, 0, time.UTC)
-	nowISO := eventlog.FormatJavaScriptISOString(now)
-
-	projectID := "project_quarantine_sync"
-	if err := repos.Projects.Upsert(ctx, storage.ProjectRecord{
-		ID: projectID, Name: "Quarantine Sync", RepoPath: filepath.Join(root, "repo"),
-		CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Projects.Upsert() error = %v", err)
-	}
-	loopID := "loop_quarantine_sync"
-	target := projectID
-	if err := repos.Loops.Upsert(ctx, storage.LoopRecord{
-		ID: loopID, Seq: 710, ProjectID: projectID, Type: "worker",
-		TargetType: "project", TargetID: &target, Status: "running",
-		CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Loops.Upsert() error = %v", err)
-	}
-	runID := "run_quarantine_sync"
-	if err := repos.Runs.Upsert(ctx, storage.RunRecord{
-		ID: runID, LoopID: loopID, Status: "running",
-		StartedAt: nowISO, CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Runs.Upsert() error = %v", err)
-	}
-	if err := repos.Queue.Upsert(ctx, storage.QueueItemRecord{
-		ID: "queue_quarantine_sync", ProjectID: &projectID, LoopID: &loopID,
-		Type: "worker", TargetType: "project", TargetID: target,
-		DedupeKey: "worker:quarantine_sync", Priority: storage.QueuePriorityWorker,
-		Status: "running", AvailableAt: nowISO, Attempts: 1, MaxAttempts: 3,
-		CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("Queue.Upsert() error = %v", err)
-	}
-	execID := "exec_quarantine_sync"
-	if err := repos.AgentExecutions.Upsert(ctx, storage.AgentExecutionRecord{
-		ID: execID, ProjectID: &projectID, LoopID: &loopID, RunID: &runID,
-		Vendor: "codex", Status: "running", StartedAt: nowISO,
-		CreatedAt: nowISO, UpdatedAt: nowISO,
-	}); err != nil {
-		t.Fatalf("AgentExecutions.Upsert() error = %v", err)
-	}
-
-	cfg, err := config.DefaultConfig(root)
-	if err != nil {
-		t.Fatalf("DefaultConfig() error = %v", err)
-	}
-	cfg.Notifications.InApp = true
-	cfg.Notifications.Osascript = config.OsascriptNotificationConfig{
-		Enabled:               true,
-		SoundForLevels:        []config.NotificationSoundLevel{config.NotificationSoundLevelActionRequired},
-		ThrottleWindowSeconds: 60,
-	}
-	osascriptPath := slowScript
-	cfg.Tools.OsascriptPath = &osascriptPath
-	cfg.Daemon.LogDir = filepath.Join(root, "logs")
-
-	rt := New(Options{Config: cfg, Logger: &testLogger{}, Now: func() time.Time { return now }})
-	execution, err := repos.AgentExecutions.GetByID(ctx, execID)
-	if err != nil || execution == nil {
-		t.Fatalf("AgentExecutions.GetByID() = %#v err=%v", execution, err)
-	}
-
-	start := time.Now()
-	quarantined, wrote, err := rt.quarantineRecoveryEvidence(ctx, repos, *execution, nowISO, "startup recovery: uncertain (test)")
-	elapsed := time.Since(start)
-	if err != nil {
-		t.Fatalf("quarantineRecoveryEvidence() error = %v", err)
-	}
-	if !quarantined || !wrote {
-		t.Fatalf("quarantineRecoveryEvidence() quarantined=%v wrote=%v, want true/true", quarantined, wrote)
-	}
-	// Budget well under a single interactive dialog (~30s) and command timeout (~35s).
-	if elapsed > 2*time.Second {
-		t.Fatalf("quarantineRecoveryEvidence took %v; must not run interactive notify on the critical path", elapsed)
-	}
-	// No notification rows yet — delivery is deferred to async recovery rescan.
-	assertHumanAttentionInAppCount(t, repos, loopID, 0)
-}
